api/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
flask>=3.0.0
